refactor: relocate constants and judge logic to the src directory to improve module structure
